Rename package `paymentsservice` to use the singular form `paymentservice`.
