+++ src/main/java/org/stellar/anchor/paymentservice/circle/model/response/CircleAccountBalancesResponse.java
@@ -1,7 +1,7 @@
-package org.stellar.anchor.paymentsservice.circle.model.response;
+package org.stellar.anchor.paymentservice.circle.model.response;
 
 import lombok.Data;
-import org.stellar.anchor.paymentsservice.circle.model.CircleBalance;
+import org.stellar.anchor.paymentservice.circle.model.CircleBalance;
 
 import java.util.List;
