+++ src/main/java/org/stellar/anchor/paymentservice/circle/model/response/CircleWalletResponse.java
@@ -1,0 +1,9 @@
+package org.stellar.anchor.paymentservice.circle.model.response;
+
+import lombok.Data;
+import org.stellar.anchor.paymentservice.circle.model.CircleWallet;
+
+@Data
+public class CircleWalletResponse {
+    CircleWallet data;
+}
